Fix: Increase speed of execution
To prevent the problem of build static- or dynamic-linked library that
is needed to link against application, it is quite difficult to
understand what is build.

CMAKE is now linking against the static version of a library and when it
is not provided it changes to the shared version. The main reason is: A
shared library must be installed on a device and the static version is
not build. But this mechanism (static or shared?) is quite difficult to
control. Because in a sysroot the library provider offers both solutions
and you can decide afterwards to install the library shared or not.